dom/devicestorage/test/test_enumerate.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html> <!--
https://bugzilla.mozilla.org/show_bug.cgi?id=717103
-->
<head>
  <title>Test for the device storage API </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="devicestorage_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=717103">Mozilla Bug 717103</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

devicestorage_setup(function () {

function enumerateSuccess(e) {

  if (e.target.result == null) {
    ok(files.length == 0, "when the enumeration is done, we shouldn't have any files in this array")
    dump("We still have length = " + files.length + "\n");
    SimpleTest.finish();
    return;
  }

  var filename = e.target.result.name;
  if (filename[0] == "/") {
    // We got /storageName/prefix/filename
    // Remove the storageName (this shows up on FirefoxOS)
    filename = filename.substring(1); // Remove leading slash
    var slashIndex = filename.indexOf("/");
    if (slashIndex >= 0) {
      filename = filename.substring(slashIndex + 1); // Remove storageName
    }
  }
  if (filename.startsWith(prefix)) {
    filename = filename.substring(prefix.length + 1); // Remove prefix
  }

  var index = files.indexOf(filename);
  files.remove(index);

  ok(index > -1, "filename should be in the enumeration : " + e.target.result.name);

  // clean up
  var cleanup = storage.delete(e.target.result.name);
  cleanup.onsuccess = function(e) {}  // todo - can i remove this?

  e.target.continue();
}

function handleError(e) {
  ok(false, "handleError was called : " + e.target.error.name);
  SimpleTest.finish();
}

function addSuccess(e) {
  addedSoFar = addedSoFar + 1;
  if (addedSoFar == files.length) {

    var cursor = storage.enumerate(prefix);
    cursor.onsuccess = enumerateSuccess;
    cursor.onerror = handleError;
  }
}

function addError(e) {
  ok(false, "addError was called : " + e.target.error.name);
  SimpleTest.finish();
}

var storage = navigator.getDeviceStorage("pictures");
ok(navigator.getDeviceStorage, "Should have getDeviceStorage");
var prefix = "devicestorage/" + randomFilename(12) + ".png"

var files = [ "a.PNG", "b.pnG", "c.png", "d/a.png", "d/b.png", "d/c.png", "d/d.png", "The/quick/brown/fox/jumps/over/the/lazy/dog.png"]
var addedSoFar = 0;


for (var i=0; i<files.length; i++) {

 request = storage.addNamed(createRandomBlob('image/png'), prefix + '/' + files[i]);

 ok(request, "Should have a non-null request");
 request.onsuccess = addSuccess;
 request.onerror = addError;
}

});

</script>
</pre>
</body>
</html>
